requirements.runtime.txt
# Runtime dependencies for Fly.io dashboard
# NO ML/training libraries - keeps image small

streamlit>=1.28.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
alpaca-trade-api>=3.1.1
schedule>=1.2.0
colorama>=0.4.6
sqlalchemy>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3
plotly>=5.18.0

